requirements.txt
kagglehub==0.3.4
pandas==2.2.3
python-dotenv==1.0.1
supabase==2.9.1
numpy==2.0.2